Pull some settings styles into their own files
