+++ src/components/scenes/PasswordRecoveryScene.js
@@ -2,9 +2,9 @@
 
 import { PasswordRecoveryScreen } from 'edge-login-ui-rn'
 import React, { Component } from 'react'
-import { View } from 'react-native'
+import { ScrollView, View } from 'react-native'
 
-import styles from '../../styles/scenes/SettingsStyle.js'
+import { THEME } from '../../theme/variables/airbitz'
 import { scale } from '../../util/scaling.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 
@@ -18,14 +18,16 @@ export default class PasswordRecovery extends Component<Props> {
   render () {
     return (
       <SceneWrapper hasTabs={false} background="body">
-        <PasswordRecoveryScreen
-          account={this.props.account}
-          context={this.props.context}
-          onComplete={this.props.onComplete}
-          onCancel={this.props.onComplete}
-          showHeader={this.props.showHeader}
-        />
-        <View style={[styles.bottomShim, { height: scale(150) }]} />
+        <ScrollView keyboardShouldPersistTaps={'always'}>
+          <PasswordRecoveryScreen
+            account={this.props.account}
+            context={this.props.context}
+            onComplete={this.props.onComplete}
+            onCancel={this.props.onComplete}
+            showHeader={this.props.showHeader}
+          />
+          <View style={{ backgroundColor: THEME.COLORS.WHITE, height: scale(150) }} />
+        </ScrollView>
       </SceneWrapper>
     )
   }
