+++ src/components/scenes/SwapSettingsScene.js
@@ -1,8 +1,8 @@
 // @flow
 
 import { type EdgePluginMap, type EdgeSwapConfig } from 'edge-core-js/types'
-import React, { Component } from 'react'
-import { Image, ScrollView, View } from 'react-native'
+import React, { type Node, Component, Fragment } from 'react'
+import { Image, ScrollView, StyleSheet, Text, View } from 'react-native'
 import CookieManager from 'react-native-cookies'
 import { Actions } from 'react-native-router-flux'
 import { connect } from 'react-redux'
@@ -11,8 +11,8 @@ import { deactivateShapeShift } from '../../actions/ShapeShiftActions.js'
 import { getSwapPluginIcon } from '../../assets/images/exchange'
 import * as Constants from '../../constants/indexConstants.js'
 import s from '../../locales/strings.js'
-import T from '../../modules/UI/components/FormattedText/index'
-import styles from '../../styles/scenes/SettingsStyle'
+import { dayText } from '../../styles/common/textStyles.js'
+import { THEME } from '../../theme/variables/airbitz.js'
 import { type Dispatch, type State as ReduxState } from '../../types/reduxTypes.js'
 import SwitchRow from '../common/RowSwitch.js'
 import { RowWithButton } from '../common/RowWithButton.js'
@@ -87,7 +87,7 @@ export class SwapSettings extends Component<Props, State> {
       <SceneWrapper hasTabs={false} background="body">
         <ScrollView>
           <View style={styles.instructionArea}>
-            <T style={styles.instructionText}>{s.strings.settings_exchange_instruction}</T>
+            <Text style={styles.instructionText}>{s.strings.settings_exchange_instruction}</Text>
           </View>
           {this.sortedIds.map(pluginId => this.renderPlugin(pluginId))}
         </ScrollView>
@@ -98,17 +98,17 @@ export class SwapSettings extends Component<Props, State> {
   renderPlugin (pluginId: string) {
     const { exchanges } = this.props
     const { displayName } = exchanges[pluginId].swapInfo
-    const logoSource = getSwapPluginIcon(pluginId)
-    const logo = <Image resizeMode={'contain'} style={styles.settingsRowLeftLogo} source={logoSource} />
+    const logo = this.renderPluginIcon(pluginId)
 
     if (pluginId === 'shapeshift') {
+      const logoSource = getSwapPluginIcon(pluginId)
       const ssLoginText = this.state.needsActivation.shapeshift ? s.strings.ss_login : s.strings.ss_logout
 
       return (
-        <View style={styles.doubleRowContainer} key={pluginId}>
+        <Fragment>
           <SwitchRow logo={logo} leftText={displayName} onToggle={() => this._onToggleEnableSwap(pluginId)} value={this.state.enabled[pluginId]} />
           <RowWithButton logo={logoSource} leftText={displayName + ' ' + s.strings.account} rightText={ssLoginText} onPress={this.shapeShiftSignInToggle} />
-        </View>
+        </Fragment>
       )
     }
 
@@ -116,7 +116,31 @@ export class SwapSettings extends Component<Props, State> {
       <SwitchRow logo={logo} key={pluginId} leftText={displayName} onToggle={() => this._onToggleEnableSwap(pluginId)} value={this.state.enabled[pluginId]} />
     )
   }
+
+  renderPluginIcon (pluginId: string): Node {
+    const logoSource = getSwapPluginIcon(pluginId)
+    return <Image resizeMode="contain" style={styles.swapIcon} source={logoSource} />
+  }
 }
+
+const iconSize = THEME.rem(1.375)
+
+const rawStyles = {
+  instructionArea: {
+    backgroundColor: THEME.COLORS.GRAY_4,
+    padding: THEME.rem(1)
+  },
+  instructionText: {
+    ...dayText('center'),
+    color: THEME.COLORS.GRAY_1
+  },
+  swapIcon: {
+    height: iconSize,
+    width: iconSize,
+    marginRight: THEME.rem(1)
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
 
 export const SwapSettingsScene = connect(
   (state: ReduxState) => ({
